cleaning up code after implementing test component for serial queries
