+++ kmaps-app/src/hooks/useSolr.js
@@ -2,8 +2,12 @@ import React from 'react';
 import { useQuery, queryCache } from 'react-query';
 import axios from 'axios';
 import jsonpAdapter from '../logic/axios-jsonp';
+// import { ReactQueryDevtools } from 'react-query-devtools';
 
-const solrurls = getSolrUrls(process.env.NODE_ENV);
+const solrurls = {
+    assets: process.env.REACT_APP_SOLR_KMASSETS + '/select',
+    terms: process.env.REACT_APP_SOLR_KMTERMS + '/select',
+};
 
 /**
  * A async function to perform a solr query provided a query object. The query object needs to have the following properties:
@@ -69,15 +73,51 @@ export function useSolrEnabled(qkey, queryobj, depvar) {
 }
 
 /**
- * A function to get the proper Solr base URL for the current environment.
- * Returns an object with both an assets and a terms property that contains the base url to that index
- * for the given environment. Uses environment variables set by .env files for each environment
- * @param env
- * @returns {{assets: string, terms: string}}
+ * Example of chained, dependent queries with the useSolr() and useSolrEnabled() hook.
+ * Finds record for place with header "Lhasa" and ADM3, then with its kmap ID finds all
+ * assets tagged with that place.
+ *
+ * @param props
+ * @returns {JSX.Element}
+ * @constructor
  */
-function getSolrUrls(env) {
-    return {
-        assets: process.env.REACT_APP_SOLR_KMASSETS + '/select',
-        terms: process.env.REACT_APP_SOLR_KMTERMS + '/select',
+function TestComp(props) {
+    const testquery = {
+        index: 'terms',
+        params: {
+            q: 'header:Lhasa AND tree:places AND feature_types:ADM3',
+            rows: 1,
+        },
     };
+    const res1 = useSolr('tqry', testquery);
+    let { numFound, start, docs } = res1
+        ? res1
+        : { numFound: 0, start: 0, docs: [] };
+    const kmapid = docs && docs.length > 0 ? docs[0].id : false;
+
+    const q2 = {
+        index: 'assets',
+        params: {
+            q: 'kmapid:' + kmapid,
+            fl: 'id,uid',
+            rows: 25,
+        },
+    };
+    const results = useSolrEnabled('newtsq', q2, res1);
+
+    let nf = '?';
+    let adocs = [];
+    if (results) {
+        console.log('res data', results);
+        nf = results.numFound;
+        adocs = JSON.stringify(results.docs, null, 2);
+    }
+    return (
+        <>
+            <p>
+                Lhasa's kmid is: {kmapid}. Found {nf} assets connected with it:{' '}
+            </p>
+            <pre>{adocs}</pre>
+        </>
+    );
 }
